requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
pdfminer.six>=20221105
pathlib
uuid 